- Split `customizable-e2e-test/browser.spec.ts` into `AAD` and `B2C` tests to mitigate concurrency. - Fix port number for `onPageLoad` test.
